// src/main/resources/angular/src/app/api/model/guild.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: rbarbado@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Character } from './character';


export interface Guild { 
    gremiId: number;
    name: string;
    owerName: string;
    email: string;
    dumGranted: Character;
    parkingBsmGranted: Character;
    expirationDate: string;
    createdDate: string;
    gremiDesc: string;
    nif: string;
    zipCode: number;
    address: string;
    cityId: number;
    telefoneNumber: number;
    state: number;
    deletedDate:string;
}